refactor: update for linter
